azure-function-simple/requirements.txt
azure-functions==1.18.0
azure-search-documents==11.4.0
azure-identity==1.15.0
openai==1.30.0
requests==2.31.0